feat: implemented CreateOrderItemUseCase in OrderItemController

+++ src/main/java/igor/henrique/dtos/order/output/OrderOutputDTO.java
@@ -8,6 +8,6 @@ public record OrderOutputDTO(
         String waiterName,
         String orderStatus,
         String orderDateTime,
-        Double totalPrice
+        BigDecimal totalPrice
 ) {
 }
